Register dark mode plugin outside of resolveConfig code path (#2368)
