+++ resolveConfig.js
@@ -2,10 +2,7 @@ const resolveConfigObjects = require('./lib/util/resolveConfig').default
 const getAllConfigs = require('./lib/util/getAllConfigs').default
 
 module.exports = function resolveConfig(...configs) {
-  // Make sure the correct config object is mutated to include flagged config plugins.
-  // This sucks, refactor soon.
-  const firstConfigWithPlugins = configs.find(c => Array.isArray(c.plugins)) || configs[0]
-  const [, ...defaultConfigs] = getAllConfigs(firstConfigWithPlugins)
+  const [, ...defaultConfigs] = getAllConfigs(configs[0])
 
   return resolveConfigObjects([...configs, ...defaultConfigs])
 }
